randoms float32

+++ statistics/stattools.py
@@ -9,21 +9,25 @@ def make_randoms(ra, dec, redshift,
         np.sin(np.deg2rad(dec.min())), 
         np.sin(np.deg2rad(dec.max())), 
         size_random
-    )
+    ).astype(np.float32)
     dec_rand = np.rad2deg(np.arcsin(sindec_rand))
     
     ra_rand  = rng.uniform(
         ra.min(), 
         ra.max(), 
         size_random
-    )
+    ).astype(np.float32)
 
     y, x  = np.histogram(redshift, 25)
     x  = 0.5*(x[1:]+x[:-1]) # center of bin
     
     ## segun numpy mejor usar la clase numpy.polynomial.Polynomial instead of np.poly1d
     poly = np.polyfit(x, y, 3)
-    zr = rng.uniform(redshift.min(), redshift.max(), size_random)
+    zr = rng.uniform(
+        redshift.min(), 
+        redshift.max(), 
+        size_random
+    ).astype(np.float32)
     poly_y = np.poly1d(poly)(zr)
     poly_y[poly_y<0] = 0.
     peso = poly_y/sum(poly_y)
